pyproject.toml
[project]
name = "instagram_content_engine"
version = "0.1.0"
description = "Instagram Content Engine using crewAI"
requires-python = ">=3.10,<3.12"
dependencies = [
    "crewai>=0.152.0",
    "python-dotenv==1.0.0",
    "opencv-python==4.8.1.78",
    "google-search-results==2.4.2",
    "unstructured>=0.18.11",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "google-generativeai>=0.8.6",
]

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']
